3.architecture_pour_la_prediction/3.4.prediction_cloud_bigqueryml_arima/requirements.txt
#pandas==0.25.2
#pyarrow==0.16.0
google-cloud-bigquery
google-api-core